Move *PickListener interfaces into separate files
