+++ platforms/android/demo/src/main/java/com/mapzen/tangram/android/MainActivity.java
@@ -10,16 +10,16 @@ import android.widget.Toast;
 
 import com.mapzen.tangram.CameraPosition;
 import com.mapzen.tangram.CameraUpdateFactory;
+import com.mapzen.tangram.FeaturePickListener;
+import com.mapzen.tangram.LabelPickListener;
 import com.mapzen.tangram.LabelPickResult;
 import com.mapzen.tangram.LngLat;
-import com.mapzen.tangram.MapController;
-import com.mapzen.tangram.MapController.FeaturePickListener;
-import com.mapzen.tangram.MapController.LabelPickListener;
-import com.mapzen.tangram.MapController.MarkerPickListener;
 import com.mapzen.tangram.MapChangeListener;
+import com.mapzen.tangram.MapController;
 import com.mapzen.tangram.MapData;
 import com.mapzen.tangram.MapView;
 import com.mapzen.tangram.Marker;
+import com.mapzen.tangram.MarkerPickListener;
 import com.mapzen.tangram.MarkerPickResult;
 import com.mapzen.tangram.SceneError;
 import com.mapzen.tangram.SceneUpdate;
